% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseabar2.R
\name{gseabar2}
\alias{gseabar2}
\title{gseabar2}
\usage{
gseabar2(
  object,
  n = 12,
  font.size = 12,
  title = "",
  label_length = 40,
  color = c("#2874C5", "#f87669")
)
}
\arguments{
\item{object}{GSEA enrich results.}

\item{n}{number of categories to show}

\item{font.size}{font size}

\item{title}{plot title}

\item{label_length}{a numeric value sets wrap length, alternatively a
custom function to format axis labels.
by default wraps names longer that 40 characters}

\item{color}{color for bar plot}

\item{...}{other parameter, ignored}
}
\value{
ggplot2 object
}
\description{
gseabar2
}
\examples{
\donttest{
# load data
test_data <- system.file("extdata", "gseares.RDS", package = "GSEAbar")
gseares <- readRDS(test_data)

# bar plot for GSEA
gseabar2(object,
       n=12,
       font.size=12,
       title="",
       label_length=40)
}
}
